Add ability to pick directories with useDirectoryPicker hook
