+++ package.json
@@ -83,10 +83,10 @@
     "react-is": "^17.0.1",
     "size-limit": "^4.9.1",
     "tsdx": "^0.14.1",
-    "tslib": "^2.0.3",
     "typescript": "3.9.7"
   },
   "dependencies": {
-    "file-selector": "0.2.4"
+    "file-selector": "0.2.4",
+    "tslib": "^2.3.1"
   }
 }
