+++ src/index.tsx
@@ -1,2 +1,4 @@
-export { default as useFilePicker } from './useFilePicker';
+import useFilePicker from './useFilePicker';
+
+export { useFilePicker as default };
 export { Validator } from './validators/validatorInterface';
